experiments/backend_next/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0

marinabox==0.1.9
langgraph==0.2.60
langchain-anthropic==0.3.1
langchain-core==0.3.28

python-dotenv>=1.0.1
aiofiles>=23.2.1